+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [0.0.43] - Big changes
+
+- `email` and `password` of `loginOrRegister` are now optional.
+- `userPublicData` has been renamed to `getUserPublicData`.
+- `publicData` has been renamed to `userPublicData`.
+
 ## [0.0.42] - Storage related code update
 
 - Storage security rules are updated and user's uid is attached to the file metadata.